Clean up a little bit
